+++ NEWS.md
@@ -2,6 +2,7 @@ MixedModels v3.1 Release Notes
 ========================
 
 * `simulate!` and thus `parametricbootstrap` methods for `GeneralizedLinearMixedModel` [#418].
+* Documented inconsistent behavior in `sdest` and `varest` `GeneralizedLinearMixedModel` [#418].
 
 MixedModels v3.0 Release Notes
 ========================